Refactor: Split cli/__init__.py into separate command modules
Split the 2,093-line CLI module into:
- parsers.py (866 lines): All argparse subparser registration
- dispatch.py (998 lines): Command dispatch and handler functions
- __init__.py (213 lines): Entry point and legacy standalone functions

The refactored __init__.py is now under the 300 line target.
All CLI functionality remains unchanged.
