+++ src/kicad_tools/cli/__init__.py
@@ -69,13 +69,17 @@ Examples:
     kct interactive --project myboard.kicad_pro
 """
 
+from __future__ import annotations
+
 import argparse
 import sys
 import traceback
-from pathlib import Path
 
 from kicad_tools import __version__
 from kicad_tools.exceptions import KiCadToolsError
+
+from .dispatch import dispatch_command
+from .parsers import register_all_parsers
 
 __all__ = [
     "main",
@@ -134,893 +138,8 @@ def main(argv: list[str] | None = None) -> int:
 
     subparsers = parser.add_subparsers(dest="command", help="Available commands")
 
-    # Symbols subcommand
-    symbols_parser = subparsers.add_parser("symbols", help="List symbols in a schematic")
-    symbols_parser.add_argument("schematic", help="Path to .kicad_sch file")
-    symbols_parser.add_argument("--format", choices=["table", "json", "csv"], default="table")
-    symbols_parser.add_argument("--filter", dest="pattern", help="Filter by reference")
-    symbols_parser.add_argument("--lib", dest="lib_id", help="Filter by library ID")
-    symbols_parser.add_argument("-v", "--verbose", action="store_true")
-
-    # Nets subcommand
-    nets_parser = subparsers.add_parser("nets", help="Trace nets in a schematic")
-    nets_parser.add_argument("schematic", help="Path to .kicad_sch file")
-    nets_parser.add_argument("--format", choices=["table", "json"], default="table")
-    nets_parser.add_argument("--net", help="Trace a specific net by label")
-    nets_parser.add_argument("--stats", action="store_true", help="Show statistics only")
-
-    # Netlist subcommand (with subcommands)
-    netlist_parser = subparsers.add_parser("netlist", help="Netlist analysis and comparison tools")
-    netlist_subparsers = netlist_parser.add_subparsers(
-        dest="netlist_command", help="Netlist commands"
-    )
-
-    # netlist analyze
-    netlist_analyze = netlist_subparsers.add_parser("analyze", help="Show connectivity statistics")
-    netlist_analyze.add_argument("netlist_schematic", help="Path to .kicad_sch file")
-    netlist_analyze.add_argument(
-        "--format", dest="netlist_format", choices=["text", "json"], default="text"
-    )
-
-    # netlist list
-    netlist_list = netlist_subparsers.add_parser(
-        "list", help="List all nets with connection counts"
-    )
-    netlist_list.add_argument("netlist_schematic", help="Path to .kicad_sch file")
-    netlist_list.add_argument(
-        "--format", dest="netlist_format", choices=["table", "json"], default="table"
-    )
-    netlist_list.add_argument(
-        "--sort",
-        dest="netlist_sort",
-        choices=["name", "connections"],
-        default="connections",
-    )
-
-    # netlist show
-    netlist_show = netlist_subparsers.add_parser("show", help="Show specific net details")
-    netlist_show.add_argument("netlist_schematic", help="Path to .kicad_sch file")
-    netlist_show.add_argument("--net", dest="netlist_net", required=True, help="Net name")
-    netlist_show.add_argument(
-        "--format", dest="netlist_format", choices=["text", "json"], default="text"
-    )
-
-    # netlist check
-    netlist_check = netlist_subparsers.add_parser("check", help="Find connectivity issues")
-    netlist_check.add_argument("netlist_schematic", help="Path to .kicad_sch file")
-    netlist_check.add_argument(
-        "--format", dest="netlist_format", choices=["text", "json"], default="text"
-    )
-
-    # netlist compare
-    netlist_compare = netlist_subparsers.add_parser("compare", help="Compare two netlists")
-    netlist_compare.add_argument("netlist_old", help="Path to old .kicad_sch file")
-    netlist_compare.add_argument("netlist_new", help="Path to new .kicad_sch file")
-    netlist_compare.add_argument(
-        "--format", dest="netlist_format", choices=["text", "json"], default="text"
-    )
-
-    # netlist export
-    netlist_export = netlist_subparsers.add_parser("export", help="Export netlist file")
-    netlist_export.add_argument("netlist_schematic", help="Path to .kicad_sch file")
-    netlist_export.add_argument("-o", "--output", dest="netlist_output", help="Output path")
-    netlist_export.add_argument(
-        "--format", dest="netlist_format", choices=["kicad", "json"], default="kicad"
-    )
-
-    # ERC subcommand
-    erc_parser = subparsers.add_parser("erc", help="Parse ERC report")
-    erc_parser.add_argument("report", help="Path to ERC report (.json or .rpt)")
-    erc_parser.add_argument("--format", choices=["table", "json", "summary"], default="table")
-    erc_parser.add_argument("--errors-only", action="store_true")
-    erc_parser.add_argument("--type", dest="filter_type", help="Filter by violation type")
-    erc_parser.add_argument("--sheet", help="Filter by sheet path")
-
-    # DRC subcommand
-    drc_parser = subparsers.add_parser("drc", help="Parse DRC report")
-    drc_parser.add_argument("report", help="Path to DRC report (.json or .rpt)")
-    drc_parser.add_argument("--format", choices=["table", "json", "summary"], default="table")
-    drc_parser.add_argument("--errors-only", action="store_true")
-    drc_parser.add_argument("--type", dest="filter_type", help="Filter by violation type")
-    drc_parser.add_argument("--net", help="Filter by net name")
-    drc_parser.add_argument(
-        "--mfr",
-        choices=["jlcpcb", "pcbway", "oshpark", "seeed"],
-        help="Check against manufacturer design rules",
-    )
-    drc_parser.add_argument("--layers", type=int, default=2, help="Number of copper layers")
-
-    # BOM subcommand
-    bom_parser = subparsers.add_parser("bom", help="Generate bill of materials")
-    bom_parser.add_argument("schematic", help="Path to .kicad_sch file")
-    bom_parser.add_argument("--format", choices=["table", "csv", "json"], default="table")
-    bom_parser.add_argument("--group", action="store_true", help="Group identical components")
-    bom_parser.add_argument(
-        "--exclude", action="append", default=[], help="Exclude references matching pattern"
-    )
-    bom_parser.add_argument("--include-dnp", action="store_true")
-    bom_parser.add_argument(
-        "--sort", choices=["reference", "value", "footprint"], default="reference"
-    )
-
-    # CHECK subcommand - pure Python DRC
-    check_parser = subparsers.add_parser("check", help="Pure Python DRC (no kicad-cli)")
-    check_parser.add_argument("pcb", help="Path to .kicad_pcb file")
-    check_parser.add_argument("--format", choices=["table", "json", "summary"], default="table")
-    check_parser.add_argument("--errors-only", action="store_true")
-    check_parser.add_argument("--strict", action="store_true", help="Exit with code 2 on warnings")
-    check_parser.add_argument(
-        "--mfr",
-        "-m",
-        choices=["jlcpcb", "pcbway", "oshpark", "seeed"],
-        default="jlcpcb",
-        help="Target manufacturer (default: jlcpcb)",
-    )
-    check_parser.add_argument("--layers", "-l", type=int, default=2, help="Number of layers")
-    check_parser.add_argument("--copper", "-c", type=float, default=1.0, help="Copper weight (oz)")
-    check_parser.add_argument(
-        "--only",
-        dest="only_checks",
-        help="Run only specific checks (comma-separated: clearance, dimensions, edge, silkscreen)",
-    )
-    check_parser.add_argument(
-        "--skip",
-        dest="skip_checks",
-        help="Skip specific checks (comma-separated)",
-    )
-    check_parser.add_argument("-v", "--verbose", action="store_true")
-
-    # SCH subcommand - schematic tools
-    sch_parser = subparsers.add_parser("sch", help="Schematic analysis tools")
-    sch_subparsers = sch_parser.add_subparsers(dest="sch_command", help="Schematic commands")
-
-    # sch summary
-    sch_summary = sch_subparsers.add_parser("summary", help="Quick schematic overview")
-    sch_summary.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_summary.add_argument("--format", choices=["text", "json"], default="text")
-    sch_summary.add_argument("-v", "--verbose", action="store_true")
-
-    # sch hierarchy
-    sch_hierarchy = sch_subparsers.add_parser("hierarchy", help="Show hierarchy tree")
-    sch_hierarchy.add_argument("schematic", help="Path to root .kicad_sch file")
-    sch_hierarchy.add_argument("--format", choices=["tree", "json"], default="tree")
-    sch_hierarchy.add_argument("--depth", type=int, help="Maximum depth to show")
-
-    # sch labels
-    sch_labels = sch_subparsers.add_parser("labels", help="List labels")
-    sch_labels.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_labels.add_argument("--format", choices=["table", "json", "csv"], default="table")
-    sch_labels.add_argument(
-        "--type", choices=["all", "local", "global", "hierarchical", "power"], default="all"
-    )
-    sch_labels.add_argument("--filter", dest="pattern", help="Filter by label text pattern")
-
-    # sch validate
-    sch_validate = sch_subparsers.add_parser("validate", help="Run validation checks")
-    sch_validate.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_validate.add_argument("--format", choices=["text", "json"], default="text")
-    sch_validate.add_argument("--strict", action="store_true", help="Exit with error on warnings")
-    sch_validate.add_argument("-q", "--quiet", action="store_true", help="Only show errors")
-
-    # sch wires
-    sch_wires = sch_subparsers.add_parser("wires", help="List wire segments and junctions")
-    sch_wires.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_wires.add_argument("--format", choices=["table", "json", "csv"], default="table")
-    sch_wires.add_argument("--stats", action="store_true", help="Show statistics only")
-    sch_wires.add_argument("--junctions", action="store_true", help="Include junction points")
-
-    # sch info
-    sch_info = sch_subparsers.add_parser("info", help="Show symbol details")
-    sch_info.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_info.add_argument("reference", help="Symbol reference (e.g., U1)")
-    sch_info.add_argument("--format", choices=["text", "json"], default="text")
-    sch_info.add_argument("--show-pins", action="store_true", help="Show pin details")
-    sch_info.add_argument("--show-properties", action="store_true", help="Show all properties")
-
-    # sch pins
-    sch_pins = sch_subparsers.add_parser("pins", help="Show symbol pin positions")
-    sch_pins.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_pins.add_argument("reference", help="Symbol reference (e.g., U1)")
-    sch_pins.add_argument("--lib", required=True, help="Path to symbol library file")
-    sch_pins.add_argument("--format", choices=["table", "json"], default="table")
-
-    # sch connections
-    sch_connections = sch_subparsers.add_parser(
-        "connections", help="Check pin connections using library positions"
-    )
-    sch_connections.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_connections.add_argument(
-        "--lib-path", action="append", dest="lib_paths", help="Library search path"
-    )
-    sch_connections.add_argument(
-        "--lib", action="append", dest="libs", help="Specific library file"
-    )
-    sch_connections.add_argument("--format", choices=["table", "json"], default="table")
-    sch_connections.add_argument("--filter", dest="pattern", help="Filter by symbol reference")
-    sch_connections.add_argument(
-        "-v", "--verbose", action="store_true", help="Show all pins, not just unconnected"
-    )
-
-    # sch unconnected
-    sch_unconnected = sch_subparsers.add_parser(
-        "unconnected", help="Find unconnected pins and issues"
-    )
-    sch_unconnected.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_unconnected.add_argument("--format", choices=["table", "json"], default="table")
-    sch_unconnected.add_argument("--filter", dest="pattern", help="Filter by symbol reference")
-    sch_unconnected.add_argument(
-        "--include-power", action="store_true", help="Include power symbols"
-    )
-    sch_unconnected.add_argument("--include-dnp", action="store_true", help="Include DNP symbols")
-
-    # sch replace
-    sch_replace = sch_subparsers.add_parser("replace", help="Replace a symbol's library ID")
-    sch_replace.add_argument("schematic", help="Path to .kicad_sch file")
-    sch_replace.add_argument("reference", help="Symbol reference to replace (e.g., U1)")
-    sch_replace.add_argument("new_lib_id", help="New library ID (e.g., 'mylib:NewSymbol')")
-    sch_replace.add_argument("--value", help="New value for the symbol")
-    sch_replace.add_argument("--footprint", help="New footprint")
-    sch_replace.add_argument("--dry-run", action="store_true", help="Show changes without applying")
-    sch_replace.add_argument("--backup", action="store_true", help="Create backup before modifying")
-
-    # PCB subcommand - PCB tools
-    pcb_parser = subparsers.add_parser("pcb", help="PCB query tools")
-    pcb_subparsers = pcb_parser.add_subparsers(dest="pcb_command", help="PCB commands")
-
-    # pcb summary
-    pcb_summary = pcb_subparsers.add_parser("summary", help="Board summary")
-    pcb_summary.add_argument("pcb", help="Path to .kicad_pcb file")
-    pcb_summary.add_argument("--format", choices=["text", "json"], default="text")
-
-    # pcb footprints
-    pcb_footprints = pcb_subparsers.add_parser("footprints", help="List footprints")
-    pcb_footprints.add_argument("pcb", help="Path to .kicad_pcb file")
-    pcb_footprints.add_argument("--format", choices=["text", "json"], default="text")
-    pcb_footprints.add_argument("--filter", dest="pattern", help="Filter by reference pattern")
-    pcb_footprints.add_argument("--sorted", action="store_true", help="Sort output")
-
-    # pcb nets
-    pcb_nets = pcb_subparsers.add_parser("nets", help="List nets")
-    pcb_nets.add_argument("pcb", help="Path to .kicad_pcb file")
-    pcb_nets.add_argument("--format", choices=["text", "json"], default="text")
-    pcb_nets.add_argument("--filter", dest="pattern", help="Filter by net pattern")
-    pcb_nets.add_argument("--sorted", action="store_true", help="Sort output")
-
-    # pcb traces
-    pcb_traces = pcb_subparsers.add_parser("traces", help="Show trace statistics")
-    pcb_traces.add_argument("pcb", help="Path to .kicad_pcb file")
-    pcb_traces.add_argument("--format", choices=["text", "json"], default="text")
-    pcb_traces.add_argument("--layer", help="Filter by layer (e.g., F.Cu)")
-
-    # pcb stackup
-    pcb_stackup = pcb_subparsers.add_parser("stackup", help="Show layer stackup")
-    pcb_stackup.add_argument("pcb", help="Path to .kicad_pcb file")
-    pcb_stackup.add_argument("--format", choices=["text", "json"], default="text")
-
-    # LIB subcommand - library tools
-    lib_parser = subparsers.add_parser("lib", help="Symbol and footprint library tools")
-    lib_subparsers = lib_parser.add_subparsers(dest="lib_command", help="Library commands")
-
-    # lib list (new - list available KiCad libraries)
-    lib_list = lib_subparsers.add_parser("list", help="List available KiCad libraries")
-    lib_list.add_argument("--symbols", action="store_true", help="List only symbol libraries")
-    lib_list.add_argument("--footprints", action="store_true", help="List only footprint libraries")
-    lib_list.add_argument("--format", choices=["table", "json"], default="table")
-
-    # lib symbols
-    lib_symbols = lib_subparsers.add_parser("symbols", help="List symbols in library")
-    lib_symbols.add_argument("library", help="Path to .kicad_sym file")
-    lib_symbols.add_argument("--format", choices=["table", "json"], default="table")
-    lib_symbols.add_argument("--pins", action="store_true", help="Show pin details")
-
-    # lib footprints (list footprints in a .pretty directory)
-    lib_footprints = lib_subparsers.add_parser(
-        "footprints", help="List footprints in a .pretty library directory"
-    )
-    lib_footprints.add_argument("directory", help="Path to .pretty directory")
-    lib_footprints.add_argument("--format", choices=["table", "json"], default="table")
-
-    # lib footprint (show details of a single .kicad_mod file)
-    lib_footprint = lib_subparsers.add_parser("footprint", help="Show details of a footprint file")
-    lib_footprint.add_argument("file", help="Path to .kicad_mod file")
-    lib_footprint.add_argument("--format", choices=["text", "json"], default="text")
-    lib_footprint.add_argument("--pads", action="store_true", help="Show pad details")
-
-    # lib symbol-info (new - show symbol details by name)
-    lib_symbol_info = lib_subparsers.add_parser(
-        "symbol-info", help="Show details of a symbol by name"
-    )
-    lib_symbol_info.add_argument("library", help="Library name or path to .kicad_sym file")
-    lib_symbol_info.add_argument("name", help="Symbol name")
-    lib_symbol_info.add_argument("--format", choices=["text", "json"], default="text")
-    lib_symbol_info.add_argument("--pins", action="store_true", help="Show pin details")
-
-    # lib footprint-info (new - show footprint details by name)
-    lib_footprint_info = lib_subparsers.add_parser(
-        "footprint-info", help="Show details of a footprint by name"
-    )
-    lib_footprint_info.add_argument("library", help="Library name or path to .pretty directory")
-    lib_footprint_info.add_argument("name", help="Footprint name")
-    lib_footprint_info.add_argument("--format", choices=["text", "json"], default="text")
-    lib_footprint_info.add_argument("--pads", action="store_true", help="Show pad details")
-
-    # lib create-symbol-lib (placeholder - requires #85)
-    lib_create_sym = lib_subparsers.add_parser(
-        "create-symbol-lib", help="Create new symbol library (not yet implemented)"
-    )
-    lib_create_sym.add_argument("path", help="Path for new .kicad_sym file")
-
-    # lib create-footprint-lib (placeholder - requires #87)
-    lib_create_fp = lib_subparsers.add_parser(
-        "create-footprint-lib", help="Create new footprint library (not yet implemented)"
-    )
-    lib_create_fp.add_argument("path", help="Path for new .pretty directory")
-
-    # lib generate-footprint (placeholder - requires #88)
-    lib_generate = lib_subparsers.add_parser(
-        "generate-footprint", help="Generate parametric footprint (not yet implemented)"
-    )
-    lib_generate.add_argument("library", help="Target library path")
-    lib_generate.add_argument(
-        "type", choices=["soic", "qfp", "qfn", "dfn", "chip", "sot"], help="Footprint type"
-    )
-    lib_generate.add_argument("--pins", type=int, help="Number of pins")
-    lib_generate.add_argument("--pitch", type=float, help="Pin pitch in mm")
-    lib_generate.add_argument("--body-width", type=float, help="Body width in mm")
-    lib_generate.add_argument("--body-size", type=float, help="Body size (square) in mm")
-    lib_generate.add_argument("--prefix", help="Footprint name prefix")
-
-    # lib export
-    lib_export = lib_subparsers.add_parser("export", help="Export library to JSON")
-    lib_export.add_argument("path", help="Path to library file or item")
-    lib_export.add_argument("--format", choices=["json"], default="json")
-
-    # FOOTPRINT subcommand - footprint generation tools
-    footprint_parser = subparsers.add_parser("footprint", help="Footprint generation and tools")
-    footprint_subparsers = footprint_parser.add_subparsers(
-        dest="footprint_command", help="Footprint commands"
-    )
-
-    # footprint generate - delegates to footprint_generate.py for subcommands
-    # This command passes all its arguments to the footprint_generate.main() function
-    footprint_subparsers.add_parser(
-        "generate",
-        help="Generate parametric footprints",
-        add_help=False,  # Let footprint_generate handle help
-    )
-
-    # MFR subcommand - manufacturer tools
-    mfr_parser = subparsers.add_parser("mfr", help="Manufacturer tools")
-    mfr_subparsers = mfr_parser.add_subparsers(dest="mfr_command", help="Manufacturer commands")
-
-    # mfr list
-    mfr_subparsers.add_parser("list", help="List available manufacturers")
-
-    # mfr info
-    mfr_info = mfr_subparsers.add_parser("info", help="Show manufacturer details")
-    mfr_info.add_argument("manufacturer", help="Manufacturer ID (jlcpcb, seeed, etc.)")
-
-    # mfr rules
-    mfr_rules = mfr_subparsers.add_parser("rules", help="Show design rules")
-    mfr_rules.add_argument("manufacturer", help="Manufacturer ID")
-    mfr_rules.add_argument("-l", "--layers", type=int, default=4, help="Layer count")
-    mfr_rules.add_argument("-c", "--copper", type=float, default=1.0, help="Copper weight (oz)")
-
-    # mfr compare
-    mfr_compare = mfr_subparsers.add_parser("compare", help="Compare manufacturers")
-    mfr_compare.add_argument("-l", "--layers", type=int, default=4, help="Layer count")
-    mfr_compare.add_argument("-c", "--copper", type=float, default=1.0, help="Copper weight (oz)")
-
-    # mfr apply-rules
-    mfr_apply = mfr_subparsers.add_parser(
-        "apply-rules", help="Apply manufacturer design rules to project/PCB"
-    )
-    mfr_apply.add_argument("file", help="Path to .kicad_pro or .kicad_pcb file")
-    mfr_apply.add_argument("manufacturer", help="Manufacturer ID (jlcpcb, seeed, etc.)")
-    mfr_apply.add_argument("-l", "--layers", type=int, default=2, help="Layer count")
-    mfr_apply.add_argument("-c", "--copper", type=float, default=1.0, help="Copper weight (oz)")
-    mfr_apply.add_argument("-o", "--output", help="Output file (default: modify in place)")
-    mfr_apply.add_argument("--dry-run", action="store_true", help="Show changes without applying")
-
-    # mfr validate
-    mfr_validate = mfr_subparsers.add_parser(
-        "validate", help="Validate PCB against manufacturer rules"
-    )
-    mfr_validate.add_argument("file", help="Path to .kicad_pcb file")
-    mfr_validate.add_argument("manufacturer", help="Manufacturer ID (jlcpcb, seeed, etc.)")
-    mfr_validate.add_argument("-l", "--layers", type=int, default=2, help="Layer count")
-    mfr_validate.add_argument("-c", "--copper", type=float, default=1.0, help="Copper weight (oz)")
-
-    # mfr export-dru
-    mfr_export_dru = mfr_subparsers.add_parser(
-        "export-dru", help="Export manufacturer rules as KiCad DRU file"
-    )
-    mfr_export_dru.add_argument("manufacturer", help="Manufacturer ID (jlcpcb, seeed, etc.)")
-    mfr_export_dru.add_argument("-l", "--layers", type=int, default=4, help="Layer count")
-    mfr_export_dru.add_argument(
-        "-c", "--copper", type=float, default=1.0, help="Copper weight (oz)"
-    )
-    mfr_export_dru.add_argument("-o", "--output", type=str, help="Output file path")
-
-    # mfr import-dru (parse an existing .kicad_dru file)
-    mfr_import_dru = mfr_subparsers.add_parser(
-        "import-dru", help="Parse and display a KiCad design rules file"
-    )
-    mfr_import_dru.add_argument("file", help="Path to .kicad_dru file")
-    mfr_import_dru.add_argument("--format", choices=["text", "json"], default="text")
-
-    # ZONES subcommand - copper zone generation
-    zones_parser = subparsers.add_parser("zones", help="Add copper pour zones to PCB")
-    zones_subparsers = zones_parser.add_subparsers(dest="zones_command", help="Zone commands")
-
-    # zones add
-    zones_add = zones_subparsers.add_parser("add", help="Add a copper zone")
-    zones_add.add_argument("pcb", help="Path to .kicad_pcb file")
-    zones_add.add_argument("-o", "--output", help="Output file path")
-    zones_add.add_argument("--net", required=True, help="Net name (e.g., GND, +3.3V)")
-    zones_add.add_argument("--layer", required=True, help="Copper layer (e.g., B.Cu, F.Cu)")
-    zones_add.add_argument("--priority", type=int, default=0, help="Zone fill priority")
-    zones_add.add_argument("--clearance", type=float, default=0.3, help="Clearance in mm")
-    zones_add.add_argument("--thermal-gap", type=float, default=0.3, help="Thermal gap in mm")
-    zones_add.add_argument(
-        "--thermal-bridge", type=float, default=0.4, help="Thermal bridge width in mm"
-    )
-    zones_add.add_argument("--min-thickness", type=float, default=0.25, help="Min thickness in mm")
-    zones_add.add_argument("-v", "--verbose", action="store_true")
-    zones_add.add_argument("--dry-run", action="store_true")
-
-    # zones list
-    zones_list = zones_subparsers.add_parser("list", help="List existing zones")
-    zones_list.add_argument("pcb", help="Path to .kicad_pcb file")
-    zones_list.add_argument("--format", choices=["text", "json"], default="text")
-
-    # zones batch
-    zones_batch = zones_subparsers.add_parser("batch", help="Add multiple zones from spec")
-    zones_batch.add_argument("pcb", help="Path to .kicad_pcb file")
-    zones_batch.add_argument("-o", "--output", help="Output file path")
-    zones_batch.add_argument(
-        "--power-nets",
-        required=True,
-        help="Power nets spec: 'NET:LAYER,...' (e.g., 'GND:B.Cu,+3.3V:F.Cu')",
-    )
-    zones_batch.add_argument("--clearance", type=float, default=0.3, help="Clearance in mm")
-    zones_batch.add_argument("-v", "--verbose", action="store_true")
-    zones_batch.add_argument("--dry-run", action="store_true")
-
-    # ROUTE subcommand - PCB autorouting
-    route_parser = subparsers.add_parser("route", help="Autoroute a PCB")
-    route_parser.add_argument("pcb", help="Path to .kicad_pcb file")
-    route_parser.add_argument("-o", "--output", help="Output file path")
-    route_parser.add_argument(
-        "--strategy",
-        choices=["basic", "negotiated", "monte-carlo"],
-        default="negotiated",
-        help="Routing strategy (default: negotiated)",
-    )
-    route_parser.add_argument("--skip-nets", help="Comma-separated nets to skip")
-    route_parser.add_argument("--grid", type=float, default=0.25, help="Grid resolution in mm")
-    route_parser.add_argument("--trace-width", type=float, default=0.2, help="Trace width in mm")
-    route_parser.add_argument("--clearance", type=float, default=0.15, help="Clearance in mm")
-    route_parser.add_argument("--via-drill", type=float, default=0.3, help="Via drill size in mm")
-    route_parser.add_argument("--via-diameter", type=float, default=0.6, help="Via diameter in mm")
-    route_parser.add_argument("--mc-trials", type=int, default=10, help="Monte Carlo trials")
-    route_parser.add_argument("--iterations", type=int, default=15, help="Max iterations")
-    route_parser.add_argument("-v", "--verbose", action="store_true")
-    route_parser.add_argument("--dry-run", action="store_true", help="Don't write output")
-    route_parser.add_argument("-q", "--quiet", action="store_true", help="Suppress progress output")
-    route_parser.add_argument(
-        "--power-nets",
-        help="Generate zones: 'NET:LAYER,...' (e.g., 'GND:B.Cu,+3.3V:F.Cu')",
-    )
-
-    # REASON subcommand - LLM-driven PCB layout
-    reason_parser = subparsers.add_parser("reason", help="LLM-driven PCB layout reasoning")
-    reason_parser.add_argument("pcb", help="Path to .kicad_pcb file")
-    reason_parser.add_argument("-o", "--output", help="Output file path")
-    reason_parser.add_argument(
-        "--export-state",
-        action="store_true",
-        help="Export state as JSON for external LLM",
-    )
-    reason_parser.add_argument("--state-output", help="Output path for state JSON")
-    reason_parser.add_argument(
-        "--interactive",
-        action="store_true",
-        help="Run interactive reasoning loop",
-    )
-    reason_parser.add_argument(
-        "--analyze",
-        action="store_true",
-        help="Print detailed analysis of PCB state",
-    )
-    reason_parser.add_argument(
-        "--auto-route",
-        action="store_true",
-        help="Auto-route priority nets without LLM",
-    )
-    reason_parser.add_argument(
-        "--max-nets",
-        type=int,
-        default=10,
-        help="Maximum nets to auto-route (default: 10)",
-    )
-    reason_parser.add_argument("--drc", help="Path to DRC report file")
-    reason_parser.add_argument("-v", "--verbose", action="store_true")
-    reason_parser.add_argument("--dry-run", action="store_true", help="Don't write output")
-
-    # OPTIMIZE-TRACES subcommand - Trace optimization
-    optimize_parser = subparsers.add_parser("optimize-traces", help="Optimize PCB traces")
-    optimize_parser.add_argument("pcb", help="Path to .kicad_pcb file")
-    optimize_parser.add_argument("-o", "--output", help="Output file (default: modify in place)")
-    optimize_parser.add_argument("--net", help="Only optimize traces matching this net pattern")
-    optimize_parser.add_argument(
-        "--no-merge", action="store_true", help="Disable collinear merging"
-    )
-    optimize_parser.add_argument(
-        "--no-zigzag", action="store_true", help="Disable zigzag elimination"
-    )
-    optimize_parser.add_argument("--no-45", action="store_true", help="Disable 45-degree corners")
-    optimize_parser.add_argument(
-        "--chamfer-size",
-        type=float,
-        default=0.5,
-        help="45-degree chamfer size in mm (default: 0.5)",
-    )
-    optimize_parser.add_argument("-v", "--verbose", action="store_true")
-    optimize_parser.add_argument(
-        "--dry-run", action="store_true", help="Show results without writing"
-    )
-    optimize_parser.add_argument(
-        "-q", "--quiet", action="store_true", help="Suppress progress output"
-    )
-
-    # VALIDATE-FOOTPRINTS subcommand
-    validate_fp_parser = subparsers.add_parser(
-        "validate-footprints", help="Validate footprints for pad spacing issues"
-    )
-    validate_fp_parser.add_argument("pcb", help="Path to .kicad_pcb file")
-    validate_fp_parser.add_argument(
-        "--min-pad-gap",
-        type=float,
-        default=0.15,
-        help="Minimum required gap between pads in mm (default: 0.15)",
-    )
-    validate_fp_parser.add_argument(
-        "--format",
-        choices=["text", "json", "summary"],
-        default="text",
-        help="Output format (default: text)",
-    )
-    validate_fp_parser.add_argument(
-        "--errors-only",
-        action="store_true",
-        help="Only show errors, not warnings",
-    )
-    # Standard library comparison options
-    validate_fp_parser.add_argument(
-        "--compare-standard",
-        action="store_true",
-        help="Compare footprints against KiCad standard library",
-    )
-    validate_fp_parser.add_argument(
-        "--tolerance",
-        type=float,
-        default=0.05,
-        help="Tolerance for standard comparison in mm (default: 0.05)",
-    )
-    validate_fp_parser.add_argument(
-        "--kicad-library-path",
-        type=str,
-        default=None,
-        help="Override path to KiCad footprint libraries",
-    )
-
-    # FIX-FOOTPRINTS subcommand
-    fix_fp_parser = subparsers.add_parser("fix-footprints", help="Fix footprint pad spacing issues")
-    fix_fp_parser.add_argument("pcb", help="Path to .kicad_pcb file")
-    fix_fp_parser.add_argument(
-        "-o",
-        "--output",
-        help="Output file path (default: modify in place)",
-    )
-    fix_fp_parser.add_argument(
-        "--min-pad-gap",
-        type=float,
-        default=0.2,
-        help="Target gap between pads in mm (default: 0.2)",
-    )
-    fix_fp_parser.add_argument(
-        "--format",
-        choices=["text", "json", "summary"],
-        default="text",
-        help="Output format (default: text)",
-    )
-    fix_fp_parser.add_argument(
-        "--dry-run",
-        action="store_true",
-        help="Show what would be changed without applying",
-    )
-
-    # PARTS subcommand - LCSC parts lookup
-    parts_parser = subparsers.add_parser("parts", help="LCSC parts lookup and search")
-    parts_subparsers = parts_parser.add_subparsers(dest="parts_command", help="Parts commands")
-
-    # parts lookup
-    parts_lookup = parts_subparsers.add_parser("lookup", help="Look up part by LCSC number")
-    parts_lookup.add_argument("part", help="LCSC part number (e.g., C123456)")
-    parts_lookup.add_argument("--format", choices=["text", "json"], default="text")
-    parts_lookup.add_argument("--no-cache", action="store_true", help="Bypass cache")
-
-    # parts search
-    parts_search = parts_subparsers.add_parser("search", help="Search for parts")
-    parts_search.add_argument("query", help="Search query")
-    parts_search.add_argument("--format", choices=["text", "json", "table"], default="table")
-    parts_search.add_argument("--limit", type=int, default=20, help="Max results")
-    parts_search.add_argument("--in-stock", action="store_true", help="Only in-stock parts")
-    parts_search.add_argument("--basic", action="store_true", help="Only JLCPCB basic parts")
-
-    # parts cache
-    parts_cache = parts_subparsers.add_parser("cache", help="Cache management")
-    parts_cache.add_argument(
-        "cache_action",
-        nargs="?",
-        choices=["stats", "clear", "clear-expired"],
-        default="stats",
-        help="Cache action (default: stats)",
-    )
-
-    # DATASHEET subcommand - datasheet search, download, and PDF parsing
-    datasheet_parser = subparsers.add_parser(
-        "datasheet", help="Datasheet search, download, and PDF parsing"
-    )
-    datasheet_subparsers = datasheet_parser.add_subparsers(
-        dest="datasheet_command", help="Datasheet commands"
-    )
-
-    # datasheet search
-    datasheet_search = datasheet_subparsers.add_parser("search", help="Search for datasheets")
-    datasheet_search.add_argument("part", help="Part number to search for")
-    datasheet_search.add_argument("--format", choices=["text", "json"], default="text")
-    datasheet_search.add_argument("--limit", type=int, default=10, help="Max results (default: 10)")
-
-    # datasheet download
-    datasheet_download = datasheet_subparsers.add_parser("download", help="Download a datasheet")
-    datasheet_download.add_argument("part", help="Part number to download")
-    datasheet_download.add_argument("-o", "--output", help="Output directory (default: cache)")
-    datasheet_download.add_argument(
-        "--force", action="store_true", help="Force download even if cached"
-    )
-
-    # datasheet list
-    datasheet_list = datasheet_subparsers.add_parser("list", help="List cached datasheets")
-    datasheet_list.add_argument("--format", choices=["text", "json"], default="text")
-
-    # datasheet cache
-    datasheet_cache = datasheet_subparsers.add_parser("cache", help="Cache management")
-    datasheet_cache.add_argument(
-        "cache_action",
-        nargs="?",
-        choices=["stats", "clear", "clear-expired"],
-        default="stats",
-        help="Cache action (default: stats)",
-    )
-    datasheet_cache.add_argument(
-        "--older-than", type=int, help="For clear: only clear entries older than N days"
-    )
-
-    # datasheet convert (PDF parsing)
-    ds_convert = datasheet_subparsers.add_parser("convert", help="Convert PDF to markdown")
-    ds_convert.add_argument("pdf", help="Path to PDF file")
-    ds_convert.add_argument("-o", "--output", help="Output file path (default: stdout)")
-    ds_convert.add_argument("--pages", help="Page range (e.g., '1-10' or '1,2,5')")
-
-    # datasheet extract-images
-    ds_images = datasheet_subparsers.add_parser("extract-images", help="Extract images from PDF")
-    ds_images.add_argument("pdf", help="Path to PDF file")
-    ds_images.add_argument("-o", "--output", required=True, help="Output directory for images")
-    ds_images.add_argument("--pages", help="Page range (e.g., '1-10' or '1,2,5')")
-    ds_images.add_argument(
-        "--min-size", type=int, default=100, help="Minimum image dimension (default: 100)"
-    )
-    ds_images.add_argument("--format", choices=["text", "json"], default="text")
-
-    # datasheet extract-tables
-    ds_tables = datasheet_subparsers.add_parser("extract-tables", help="Extract tables from PDF")
-    ds_tables.add_argument("pdf", help="Path to PDF file")
-    ds_tables.add_argument("-o", "--output", help="Output directory for tables")
-    ds_tables.add_argument("--pages", help="Page range (e.g., '1-10' or '1,2,5')")
-    ds_tables.add_argument("--format", choices=["markdown", "csv", "json"], default="markdown")
-
-    # datasheet info
-    ds_info = datasheet_subparsers.add_parser("info", help="Show PDF information")
-    ds_info.add_argument("pdf", help="Path to PDF file")
-    ds_info.add_argument("--format", choices=["text", "json"], default="text")
-
-    # PLACEMENT subcommand - placement conflict detection and resolution
-    placement_parser = subparsers.add_parser("placement", help="Detect and fix placement conflicts")
-    placement_subparsers = placement_parser.add_subparsers(
-        dest="placement_command", help="Placement commands"
-    )
-
-    # placement check
-    placement_check = placement_subparsers.add_parser(
-        "check", help="Check PCB for placement conflicts"
-    )
-    placement_check.add_argument("pcb", help="Path to .kicad_pcb file")
-    placement_check.add_argument("--format", choices=["table", "json", "summary"], default="table")
-    placement_check.add_argument(
-        "--pad-clearance", type=float, default=0.1, help="Min pad clearance (mm)"
-    )
-    placement_check.add_argument(
-        "--hole-clearance", type=float, default=0.5, help="Min hole clearance (mm)"
-    )
-    placement_check.add_argument(
-        "--edge-clearance", type=float, default=0.3, help="Min edge clearance (mm)"
-    )
-    placement_check.add_argument("-v", "--verbose", action="store_true")
-    placement_check.add_argument(
-        "-q", "--quiet", action="store_true", help="Suppress progress output"
-    )
-
-    # placement fix
-    placement_fix = placement_subparsers.add_parser("fix", help="Suggest and apply placement fixes")
-    placement_fix.add_argument("pcb", help="Path to .kicad_pcb file")
-    placement_fix.add_argument("-o", "--output", help="Output file path")
-    placement_fix.add_argument(
-        "--strategy",
-        choices=["spread", "compact", "anchor"],
-        default="spread",
-        help="Fix strategy",
-    )
-    placement_fix.add_argument("--anchor", help="Comma-separated components to keep fixed")
-    placement_fix.add_argument("--dry-run", action="store_true")
-    placement_fix.add_argument("-v", "--verbose", action="store_true")
-    placement_fix.add_argument(
-        "-q", "--quiet", action="store_true", help="Suppress progress output"
-    )
-
-    # placement optimize
-    placement_optimize = placement_subparsers.add_parser(
-        "optimize", help="Optimize placement for routability"
-    )
-    placement_optimize.add_argument("pcb", help="Path to .kicad_pcb file")
-    placement_optimize.add_argument("-o", "--output", help="Output file path")
-    placement_optimize.add_argument(
-        "--strategy",
-        choices=["force-directed", "evolutionary", "hybrid"],
-        default="force-directed",
-        help="Optimization strategy (default: force-directed)",
-    )
-    placement_optimize.add_argument(
-        "--iterations",
-        type=int,
-        default=1000,
-        help="Max iterations for physics simulation (default: 1000)",
-    )
-    placement_optimize.add_argument(
-        "--generations",
-        type=int,
-        default=100,
-        help="Generations for evolutionary/hybrid mode (default: 100)",
-    )
-    placement_optimize.add_argument(
-        "--population",
-        type=int,
-        default=50,
-        help="Population size for evolutionary/hybrid mode (default: 50)",
-    )
-    placement_optimize.add_argument(
-        "--grid",
-        type=float,
-        default=0.0,
-        help="Position grid snap in mm (0 to disable, default: 0)",
-    )
-    placement_optimize.add_argument(
-        "--fixed",
-        help="Comma-separated component refs to keep fixed (e.g., J1,J2,H1)",
-    )
-    placement_optimize.add_argument("--dry-run", action="store_true", help="Preview only")
-    placement_optimize.add_argument("-v", "--verbose", action="store_true")
-    placement_optimize.add_argument(
-        "-q", "--quiet", action="store_true", help="Suppress progress output"
-    )
-
-    # CONFIG subcommand - configuration management
-    config_parser = subparsers.add_parser("config", help="View and manage configuration")
-    config_parser.add_argument(
-        "--show",
-        action="store_true",
-        help="Show effective configuration with sources",
-    )
-    config_parser.add_argument(
-        "--init",
-        action="store_true",
-        help="Create template config file",
-    )
-    config_parser.add_argument(
-        "--paths",
-        action="store_true",
-        help="Show config file paths",
-    )
-    config_parser.add_argument(
-        "--user",
-        action="store_true",
-        help="Use user config for --init",
-    )
-    config_parser.add_argument(
-        "config_action",
-        nargs="?",
-        choices=["get", "set"],
-        help="Config action",
-    )
-    config_parser.add_argument(
-        "config_key",
-        nargs="?",
-        help="Config key (e.g., defaults.format)",
-    )
-    config_parser.add_argument(
-        "config_value",
-        nargs="?",
-        help="Value to set",
-    )
-
-    # INTERACTIVE subcommand - REPL mode
-    interactive_parser = subparsers.add_parser("interactive", help="Launch interactive REPL mode")
-    interactive_parser.add_argument(
-        "--project",
-        help="Auto-load a project on startup",
-    )
-
-    # VALIDATE subcommand - validation tools
-    validate_parser = subparsers.add_parser("validate", help="Validation tools")
-    validate_parser.add_argument(
-        "validate_project",
-        nargs="?",
-        help="Path to .kicad_pro file",
-    )
-    validate_parser.add_argument(
-        "--sync",
-        action="store_true",
-        help="Check schematic-to-PCB netlist synchronization",
-    )
-    validate_parser.add_argument(
-        "--schematic",
-        "-s",
-        dest="validate_schematic",
-        help="Path to .kicad_sch file (if not using project file)",
-    )
-    validate_parser.add_argument(
-        "--pcb",
-        "-p",
-        dest="validate_pcb",
-        help="Path to .kicad_pcb file (if not using project file)",
-    )
-    validate_parser.add_argument(
-        "--format",
-        dest="validate_format",
-        choices=["table", "json", "summary"],
-        default="table",
-        help="Output format (default: table)",
-    )
-    validate_parser.add_argument(
-        "--errors-only",
-        dest="validate_errors_only",
-        action="store_true",
-        help="Show only errors, not warnings",
-    )
-    validate_parser.add_argument(
-        "--strict",
-        dest="validate_strict",
-        action="store_true",
-        help="Exit with error code on warnings",
-    )
-    validate_parser.add_argument(
-        "-v",
-        "--verbose",
-        dest="validate_verbose",
-        action="store_true",
-        help="Show detailed issue information",
-    )
+    # Register all subparsers from the parsers module
+    register_all_parsers(subparsers)
 
     # Handle footprint generate specially - it has its own subcommand parser
     if argv is None:
@@ -1040,7 +159,7 @@ def main(argv: list[str] | None = None) -> int:
     verbose = getattr(args, "global_verbose", False)
 
     try:
-        return _dispatch_command(args)
+        return dispatch_command(args)
     except KeyboardInterrupt:
         print("\nInterrupted", file=sys.stderr)
         return 130
@@ -1052,1006 +171,7 @@ def main(argv: list[str] | None = None) -> int:
         return 1
 
 
-def _dispatch_command(args) -> int:
-    """Dispatch to the appropriate command handler."""
-    if args.command == "symbols":
-        from .symbols import main as symbols_cmd
-
-        # Convert args back to argv for the subcommand
-        sub_argv = [args.schematic]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.pattern:
-            sub_argv.extend(["--filter", args.pattern])
-        if args.lib_id:
-            sub_argv.extend(["--lib", args.lib_id])
-        if args.verbose:
-            sub_argv.append("--verbose")
-        return symbols_cmd(sub_argv)
-
-    elif args.command == "nets":
-        from .nets import main as nets_cmd
-
-        sub_argv = [args.schematic]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.net:
-            sub_argv.extend(["--net", args.net])
-        if args.stats:
-            sub_argv.append("--stats")
-        return nets_cmd(sub_argv)
-
-    elif args.command == "netlist":
-        from .netlist_cmd import main as netlist_cmd
-
-        if not args.netlist_command:
-            # No subcommand, show help
-            return netlist_cmd(["--help"])
-
-        sub_argv = [args.netlist_command]
-
-        if args.netlist_command == "compare":
-            sub_argv.extend([args.netlist_old, args.netlist_new])
-        elif args.netlist_command == "export":
-            sub_argv.append(args.netlist_schematic)
-            if hasattr(args, "netlist_output") and args.netlist_output:
-                sub_argv.extend(["-o", args.netlist_output])
-        else:
-            sub_argv.append(args.netlist_schematic)
-
-        if hasattr(args, "netlist_format") and args.netlist_format:
-            sub_argv.extend(["--format", args.netlist_format])
-        if hasattr(args, "netlist_sort") and args.netlist_sort != "connections":
-            sub_argv.extend(["--sort", args.netlist_sort])
-        if hasattr(args, "netlist_net") and args.netlist_net:
-            sub_argv.extend(["--net", args.netlist_net])
-
-        return netlist_cmd(sub_argv)
-
-    elif args.command == "erc":
-        from .erc_cmd import main as erc_cmd
-
-        sub_argv = [args.report]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.errors_only:
-            sub_argv.append("--errors-only")
-        if args.filter_type:
-            sub_argv.extend(["--type", args.filter_type])
-        if args.sheet:
-            sub_argv.extend(["--sheet", args.sheet])
-        return erc_cmd(sub_argv)
-
-    elif args.command == "drc":
-        from .drc_cmd import main as drc_cmd
-
-        sub_argv = [args.report]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.errors_only:
-            sub_argv.append("--errors-only")
-        if args.filter_type:
-            sub_argv.extend(["--type", args.filter_type])
-        if args.net:
-            sub_argv.extend(["--net", args.net])
-        if args.mfr:
-            sub_argv.extend(["--mfr", args.mfr])
-        if args.layers != 2:
-            sub_argv.extend(["--layers", str(args.layers)])
-        return drc_cmd(sub_argv)
-
-    elif args.command == "bom":
-        from .bom_cmd import main as bom_cmd
-
-        sub_argv = [args.schematic]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.group:
-            sub_argv.append("--group")
-        for pattern in args.exclude:
-            sub_argv.extend(["--exclude", pattern])
-        if args.include_dnp:
-            sub_argv.append("--include-dnp")
-        if args.sort != "reference":
-            sub_argv.extend(["--sort", args.sort])
-        return bom_cmd(sub_argv)
-
-    elif args.command == "check":
-        return _run_check_command(args)
-
-    elif args.command == "sch":
-        return _run_sch_command(args)
-
-    elif args.command == "pcb":
-        return _run_pcb_command(args)
-
-    elif args.command == "lib":
-        return _run_lib_command(args)
-
-    elif args.command == "footprint":
-        return _run_footprint_command(args)
-
-    elif args.command == "mfr":
-        return _run_mfr_command(args)
-
-    elif args.command == "parts":
-        return _run_parts_command(args)
-
-    elif args.command == "datasheet":
-        return _run_datasheet_command(args)
-
-    elif args.command == "zones":
-        return _run_zones_command(args)
-
-    elif args.command == "route":
-        return _run_route_command(args)
-
-    elif args.command == "reason":
-        return _run_reason_command(args)
-
-    elif args.command == "placement":
-        return _run_placement_command(args)
-
-    elif args.command == "optimize-traces":
-        return _run_optimize_command(args)
-
-    elif args.command == "validate-footprints":
-        return _run_validate_footprints_command(args)
-
-    elif args.command == "fix-footprints":
-        return _run_fix_footprints_command(args)
-
-    elif args.command == "config":
-        return _run_config_command(args)
-
-    elif args.command == "interactive":
-        return _run_interactive_command(args)
-
-    elif args.command == "validate":
-        return _run_validate_command(args)
-
-    return 0
-
-
-def _run_validate_footprints_command(args) -> int:
-    """Handle validate-footprints command."""
-    from .footprint_cmd import main_validate
-
-    sub_argv = [args.pcb]
-    if args.min_pad_gap != 0.15:
-        sub_argv.extend(["--min-pad-gap", str(args.min_pad_gap)])
-    if args.format != "text":
-        sub_argv.extend(["--format", args.format])
-    if args.errors_only:
-        sub_argv.append("--errors-only")
-    # Standard comparison options
-    if getattr(args, "compare_standard", False):
-        sub_argv.append("--compare-standard")
-    if getattr(args, "tolerance", 0.05) != 0.05:
-        sub_argv.extend(["--tolerance", str(args.tolerance)])
-    if getattr(args, "kicad_library_path", None):
-        sub_argv.extend(["--kicad-library-path", args.kicad_library_path])
-    # Use global quiet flag
-    if getattr(args, "global_quiet", False):
-        sub_argv.append("--quiet")
-    return main_validate(sub_argv)
-
-
-def _run_fix_footprints_command(args) -> int:
-    """Handle fix-footprints command."""
-    from .footprint_cmd import main_fix
-
-    sub_argv = [args.pcb]
-    if args.output:
-        sub_argv.extend(["-o", args.output])
-    if args.min_pad_gap != 0.2:
-        sub_argv.extend(["--min-pad-gap", str(args.min_pad_gap)])
-    if args.format != "text":
-        sub_argv.extend(["--format", args.format])
-    if args.dry_run:
-        sub_argv.append("--dry-run")
-    # Use global quiet flag
-    if getattr(args, "global_quiet", False):
-        sub_argv.append("--quiet")
-    return main_fix(sub_argv)
-
-
-def _run_check_command(args) -> int:
-    """Handle check command (pure Python DRC)."""
-    from .check_cmd import main as check_main
-
-    sub_argv = [args.pcb]
-    if args.format != "table":
-        sub_argv.extend(["--format", args.format])
-    if args.errors_only:
-        sub_argv.append("--errors-only")
-    if args.strict:
-        sub_argv.append("--strict")
-    if args.mfr != "jlcpcb":
-        sub_argv.extend(["--mfr", args.mfr])
-    if args.layers != 2:
-        sub_argv.extend(["--layers", str(args.layers)])
-    if args.copper != 1.0:
-        sub_argv.extend(["--copper", str(args.copper)])
-    if args.only_checks:
-        sub_argv.extend(["--only", args.only_checks])
-    if args.skip_checks:
-        sub_argv.extend(["--skip", args.skip_checks])
-    if args.verbose:
-        sub_argv.append("--verbose")
-    return check_main(sub_argv)
-
-
-def _run_sch_command(args) -> int:
-    """Handle schematic subcommands."""
-    if not args.sch_command:
-        print("Usage: kicad-tools sch <command> [options] <file>")
-        print("Commands: summary, hierarchy, labels, validate, wires, info, pins,")
-        print("          connections, unconnected, replace")
-        return 1
-
-    schematic_path = Path(args.schematic)
-    if not schematic_path.exists():
-        print(f"Error: File not found: {schematic_path}", file=sys.stderr)
-        return 1
-
-    if args.sch_command == "summary":
-        from .sch_summary import run_summary
-
-        return run_summary(schematic_path, args.format, args.verbose)
-
-    elif args.sch_command == "hierarchy":
-        from .sch_hierarchy import main as hierarchy_main
-
-        sub_argv = [str(schematic_path), "tree"]
-        if args.format != "tree":
-            sub_argv.extend(["--format", args.format])
-        if args.depth:
-            sub_argv.extend(["--depth", str(args.depth)])
-        return hierarchy_main(sub_argv) or 0
-
-    elif args.sch_command == "labels":
-        from .sch_list_labels import main as labels_main
-
-        sub_argv = [str(schematic_path)]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.type != "all":
-            sub_argv.extend(["--type", args.type])
-        if args.pattern:
-            sub_argv.extend(["--filter", args.pattern])
-        return labels_main(sub_argv) or 0
-
-    elif args.sch_command == "validate":
-        from .sch_validate import main as validate_main
-
-        sub_argv = [str(schematic_path)]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        if args.strict:
-            sub_argv.append("--strict")
-        if args.quiet:
-            sub_argv.append("--quiet")
-        return validate_main(sub_argv) or 0
-
-    elif args.sch_command == "wires":
-        from .sch_list_wires import main as wires_main
-
-        sub_argv = [str(schematic_path)]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.stats:
-            sub_argv.append("--stats")
-        if args.junctions:
-            sub_argv.append("--junctions")
-        return wires_main(sub_argv) or 0
-
-    elif args.sch_command == "info":
-        from .sch_symbol_info import main as info_main
-
-        sub_argv = [str(schematic_path), args.reference]
-        if args.format == "json":
-            sub_argv.append("--json")
-        if args.show_pins:
-            sub_argv.append("--show-pins")
-        if args.show_properties:
-            sub_argv.append("--show-properties")
-        return info_main(sub_argv) or 0
-
-    elif args.sch_command == "pins":
-        from .sch_pin_positions import main as pins_main
-
-        sub_argv = [str(schematic_path), args.reference, "--lib", args.lib]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        return pins_main(sub_argv) or 0
-
-    elif args.sch_command == "connections":
-        from .sch_check_connections import main as connections_main
-
-        sub_argv = [str(schematic_path)]
-        if args.lib_paths:
-            for path in args.lib_paths:
-                sub_argv.extend(["--lib-path", path])
-        if args.libs:
-            for lib in args.libs:
-                sub_argv.extend(["--lib", lib])
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.pattern:
-            sub_argv.extend(["--filter", args.pattern])
-        if args.verbose:
-            sub_argv.append("--verbose")
-        return connections_main(sub_argv) or 0
-
-    elif args.sch_command == "unconnected":
-        from .sch_find_unconnected import main as unconnected_main
-
-        sub_argv = [str(schematic_path)]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.pattern:
-            sub_argv.extend(["--filter", args.pattern])
-        if args.include_power:
-            sub_argv.append("--include-power")
-        if args.include_dnp:
-            sub_argv.append("--include-dnp")
-        return unconnected_main(sub_argv) or 0
-
-    elif args.sch_command == "replace":
-        from .sch_replace_symbol import main as replace_main
-
-        sub_argv = [str(schematic_path), args.reference, args.new_lib_id]
-        if args.value:
-            sub_argv.extend(["--value", args.value])
-        if args.footprint:
-            sub_argv.extend(["--footprint", args.footprint])
-        if args.dry_run:
-            sub_argv.append("--dry-run")
-        if args.backup:
-            sub_argv.append("--backup")
-        return replace_main(sub_argv) or 0
-
-    return 1
-
-
-def _run_pcb_command(args) -> int:
-    """Handle PCB subcommands."""
-    if not args.pcb_command:
-        print("Usage: kicad-tools pcb <command> [options] <file>")
-        print("Commands: summary, footprints, nets, traces, stackup")
-        return 1
-
-    pcb_path = Path(args.pcb)
-    if not pcb_path.exists():
-        print(f"Error: File not found: {pcb_path}", file=sys.stderr)
-        return 1
-
-    from .pcb_query import main as pcb_main
-
-    if args.pcb_command == "summary":
-        sub_argv = [str(pcb_path), "summary"]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        return pcb_main(sub_argv) or 0
-
-    elif args.pcb_command == "footprints":
-        sub_argv = [str(pcb_path), "footprints"]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        if args.pattern:
-            sub_argv.extend(["--filter", args.pattern])
-        if args.sorted:
-            sub_argv.append("--sorted")
-        return pcb_main(sub_argv) or 0
-
-    elif args.pcb_command == "nets":
-        sub_argv = [str(pcb_path), "nets"]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        if args.pattern:
-            sub_argv.extend(["--filter", args.pattern])
-        if args.sorted:
-            sub_argv.append("--sorted")
-        return pcb_main(sub_argv) or 0
-
-    elif args.pcb_command == "traces":
-        sub_argv = [str(pcb_path), "traces"]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        if args.layer:
-            sub_argv.extend(["--layer", args.layer])
-        return pcb_main(sub_argv) or 0
-
-    elif args.pcb_command == "stackup":
-        sub_argv = [str(pcb_path), "stackup"]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        return pcb_main(sub_argv) or 0
-
-    return 1
-
-
-def _run_lib_command(args) -> int:
-    """Handle library subcommands."""
-    if not args.lib_command:
-        print("Usage: kicad-tools lib <command> [options] <file>")
-        print("Commands: list, symbols, footprints, footprint, symbol-info, footprint-info,")
-        print("          create-symbol-lib, create-footprint-lib, generate-footprint, export")
-        return 1
-
-    # Use the new unified lib module for new commands
-    from .lib import (
-        create_footprint_library,
-        create_symbol_library,
-        export_library,
-        generate_footprint,
-        list_kicad_libraries,
-        show_footprint_info,
-        show_symbol_info,
-    )
-
-    if args.lib_command == "list":
-        lib_type = "all"
-        if getattr(args, "symbols", False) and not getattr(args, "footprints", False):
-            lib_type = "symbols"
-        elif getattr(args, "footprints", False) and not getattr(args, "symbols", False):
-            lib_type = "footprints"
-        return list_kicad_libraries(lib_type, args.format)
-
-    elif args.lib_command == "symbols":
-        library_path = Path(args.library)
-        if not library_path.exists():
-            print(f"Error: File not found: {library_path}", file=sys.stderr)
-            return 1
-
-        from .lib_list_symbols import main as lib_main
-
-        sub_argv = [str(library_path)]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.pins:
-            sub_argv.append("--pins")
-        return lib_main(sub_argv) or 0
-
-    elif args.lib_command == "footprints":
-        from .lib_footprints import list_footprints
-
-        directory_path = Path(args.directory)
-        if not directory_path.exists():
-            print(f"Error: Directory not found: {directory_path}", file=sys.stderr)
-            return 1
-        return list_footprints(directory_path, args.format)
-
-    elif args.lib_command == "footprint":
-        from .lib_footprints import show_footprint
-
-        file_path = Path(args.file)
-        if not file_path.exists():
-            print(f"Error: File not found: {file_path}", file=sys.stderr)
-            return 1
-        return show_footprint(file_path, args.format, getattr(args, "pads", False))
-
-    elif args.lib_command == "symbol-info":
-        return show_symbol_info(
-            args.library,
-            args.name,
-            args.format,
-            getattr(args, "pins", False),
-        )
-
-    elif args.lib_command == "footprint-info":
-        return show_footprint_info(
-            args.library,
-            args.name,
-            args.format,
-            getattr(args, "pads", False),
-        )
-
-    elif args.lib_command == "create-symbol-lib":
-        return create_symbol_library(args.path)
-
-    elif args.lib_command == "create-footprint-lib":
-        return create_footprint_library(args.path)
-
-    elif args.lib_command == "generate-footprint":
-        return generate_footprint(
-            args.library,
-            args.type,
-            pins=getattr(args, "pins", None),
-            pitch=getattr(args, "pitch", None),
-            body_width=getattr(args, "body_width", None),
-            body_size=getattr(args, "body_size", None),
-            prefix=getattr(args, "prefix", None),
-        )
-
-    elif args.lib_command == "export":
-        return export_library(args.path, args.format)
-
-    return 1
-
-
-def _run_footprint_command(args) -> int:
-    """Handle footprint subcommands."""
-    if not args.footprint_command:
-        print("Usage: kicad-tools footprint <command> [options]")
-        print("Commands: generate")
-        return 1
-
-    if args.footprint_command == "generate":
-        from .footprint_generate import main as generate_main
-
-        # Pass all remaining arguments to the generate subcommand
-        sub_argv = args.fp_args if hasattr(args, "fp_args") and args.fp_args else []
-        return generate_main(sub_argv) or 0
-
-    return 1
-
-
-def _run_mfr_command(args) -> int:
-    """Handle manufacturer subcommands."""
-    if not args.mfr_command:
-        print("Usage: kicad-tools mfr <command> [options]")
-        print("Commands: list, info, rules, compare, apply-rules, validate, export-dru, import-dru")
-        return 1
-
-    from .mfr import main as mfr_main
-
-    if args.mfr_command == "list":
-        return mfr_main(["list"]) or 0
-
-    elif args.mfr_command == "info":
-        return mfr_main(["info", args.manufacturer]) or 0
-
-    elif args.mfr_command == "rules":
-        sub_argv = ["rules", args.manufacturer]
-        if args.layers != 4:
-            sub_argv.extend(["--layers", str(args.layers)])
-        if args.copper != 1.0:
-            sub_argv.extend(["--copper", str(args.copper)])
-        return mfr_main(sub_argv) or 0
-
-    elif args.mfr_command == "compare":
-        sub_argv = ["compare"]
-        if args.layers != 4:
-            sub_argv.extend(["--layers", str(args.layers)])
-        if args.copper != 1.0:
-            sub_argv.extend(["--copper", str(args.copper)])
-        return mfr_main(sub_argv) or 0
-
-    elif args.mfr_command == "apply-rules":
-        sub_argv = ["apply-rules", args.file, args.manufacturer]
-        if args.layers != 2:
-            sub_argv.extend(["--layers", str(args.layers)])
-        if args.copper != 1.0:
-            sub_argv.extend(["--copper", str(args.copper)])
-        if args.output:
-            sub_argv.extend(["--output", args.output])
-        if args.dry_run:
-            sub_argv.append("--dry-run")
-        return mfr_main(sub_argv) or 0
-
-    elif args.mfr_command == "validate":
-        sub_argv = ["validate", args.file, args.manufacturer]
-        if args.layers != 2:
-            sub_argv.extend(["--layers", str(args.layers)])
-        if args.copper != 1.0:
-            sub_argv.extend(["--copper", str(args.copper)])
-        return mfr_main(sub_argv) or 0
-
-    elif args.mfr_command == "export-dru":
-        sub_argv = ["export-dru", args.manufacturer]
-        if args.layers != 4:
-            sub_argv.extend(["--layers", str(args.layers)])
-        if args.copper != 1.0:
-            sub_argv.extend(["--copper", str(args.copper)])
-        if args.output:
-            sub_argv.extend(["--output", args.output])
-        return mfr_main(sub_argv) or 0
-
-    elif args.mfr_command == "import-dru":
-        from .mfr_dru import import_dru
-
-        file_path = Path(args.file)
-        if not file_path.exists():
-            print(f"Error: File not found: {file_path}", file=sys.stderr)
-            return 1
-        return import_dru(file_path, args.format)
-
-    return 1
-
-
-def _run_parts_command(args) -> int:
-    """Handle parts subcommands."""
-    if not args.parts_command:
-        print("Usage: kicad-tools parts <command> [options]")
-        print("Commands: lookup, search, cache")
-        return 1
-
-    from .parts_cmd import main as parts_main
-
-    if args.parts_command == "lookup":
-        sub_argv = ["lookup", args.part]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        if args.no_cache:
-            sub_argv.append("--no-cache")
-        return parts_main(sub_argv) or 0
-
-    elif args.parts_command == "search":
-        sub_argv = ["search", args.query]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.limit != 20:
-            sub_argv.extend(["--limit", str(args.limit)])
-        if args.in_stock:
-            sub_argv.append("--in-stock")
-        if args.basic:
-            sub_argv.append("--basic")
-        return parts_main(sub_argv) or 0
-
-    elif args.parts_command == "cache":
-        sub_argv = ["cache", args.cache_action]
-        return parts_main(sub_argv) or 0
-
-    return 1
-
-
-def _run_datasheet_command(args) -> int:
-    """Handle datasheet subcommands."""
-    if not args.datasheet_command:
-        print("Usage: kicad-tools datasheet <command> [options]")
-        print(
-            "Commands: search, download, list, cache, convert, extract-images, extract-tables, info"
-        )
-        return 1
-
-    from .datasheet_cmd import main as datasheet_main
-
-    if args.datasheet_command == "search":
-        sub_argv = ["search", args.part]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        if args.limit != 10:
-            sub_argv.extend(["--limit", str(args.limit)])
-        return datasheet_main(sub_argv) or 0
-
-    elif args.datasheet_command == "download":
-        sub_argv = ["download", args.part]
-        if args.output:
-            sub_argv.extend(["-o", args.output])
-        if args.force:
-            sub_argv.append("--force")
-        return datasheet_main(sub_argv) or 0
-
-    elif args.datasheet_command == "list":
-        sub_argv = ["list"]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        return datasheet_main(sub_argv) or 0
-
-    elif args.datasheet_command == "cache":
-        sub_argv = ["cache", args.cache_action]
-        if getattr(args, "older_than", None):
-            sub_argv.extend(["--older-than", str(args.older_than)])
-        return datasheet_main(sub_argv) or 0
-
-    elif args.datasheet_command == "convert":
-        sub_argv = ["convert", args.pdf]
-        if args.output:
-            sub_argv.extend(["-o", args.output])
-        if args.pages:
-            sub_argv.extend(["--pages", args.pages])
-        return datasheet_main(sub_argv) or 0
-
-    elif args.datasheet_command == "extract-images":
-        sub_argv = ["extract-images", args.pdf, "-o", args.output]
-        if args.pages:
-            sub_argv.extend(["--pages", args.pages])
-        if args.min_size != 100:
-            sub_argv.extend(["--min-size", str(args.min_size)])
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        return datasheet_main(sub_argv) or 0
-
-    elif args.datasheet_command == "extract-tables":
-        sub_argv = ["extract-tables", args.pdf]
-        if args.output:
-            sub_argv.extend(["-o", args.output])
-        if args.pages:
-            sub_argv.extend(["--pages", args.pages])
-        if args.format != "markdown":
-            sub_argv.extend(["--format", args.format])
-        return datasheet_main(sub_argv) or 0
-
-    elif args.datasheet_command == "info":
-        sub_argv = ["info", args.pdf]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        return datasheet_main(sub_argv) or 0
-
-    return 1
-
-
-def _run_zones_command(args) -> int:
-    """Handle zones command."""
-    if not args.zones_command:
-        print("Usage: kicad-tools zones <command> [options] <file>")
-        print("Commands: add, list, batch")
-        return 1
-
-    from .zones_cmd import main as zones_main
-
-    if args.zones_command == "add":
-        sub_argv = ["add", args.pcb]
-        if args.output:
-            sub_argv.extend(["-o", args.output])
-        sub_argv.extend(["--net", args.net])
-        sub_argv.extend(["--layer", args.layer])
-        if args.priority != 0:
-            sub_argv.extend(["--priority", str(args.priority)])
-        if args.clearance != 0.3:
-            sub_argv.extend(["--clearance", str(args.clearance)])
-        if getattr(args, "thermal_gap", 0.3) != 0.3:
-            sub_argv.extend(["--thermal-gap", str(args.thermal_gap)])
-        if getattr(args, "thermal_bridge", 0.4) != 0.4:
-            sub_argv.extend(["--thermal-bridge", str(args.thermal_bridge)])
-        if getattr(args, "min_thickness", 0.25) != 0.25:
-            sub_argv.extend(["--min-thickness", str(args.min_thickness)])
-        if args.verbose:
-            sub_argv.append("--verbose")
-        if args.dry_run:
-            sub_argv.append("--dry-run")
-        # Use global quiet flag
-        if getattr(args, "global_quiet", False):
-            sub_argv.append("--quiet")
-        return zones_main(sub_argv) or 0
-
-    elif args.zones_command == "list":
-        sub_argv = ["list", args.pcb]
-        if args.format != "text":
-            sub_argv.extend(["--format", args.format])
-        return zones_main(sub_argv) or 0
-
-    elif args.zones_command == "batch":
-        sub_argv = ["batch", args.pcb]
-        if args.output:
-            sub_argv.extend(["-o", args.output])
-        sub_argv.extend(["--power-nets", args.power_nets])
-        if args.clearance != 0.3:
-            sub_argv.extend(["--clearance", str(args.clearance)])
-        if args.verbose:
-            sub_argv.append("--verbose")
-        if args.dry_run:
-            sub_argv.append("--dry-run")
-        # Use global quiet flag
-        if getattr(args, "global_quiet", False):
-            sub_argv.append("--quiet")
-        return zones_main(sub_argv) or 0
-
-    return 1
-
-
-def _run_route_command(args) -> int:
-    """Handle route command."""
-    from .route_cmd import main as route_main
-
-    sub_argv = [args.pcb]
-    if args.output:
-        sub_argv.extend(["-o", args.output])
-    if args.strategy != "negotiated":
-        sub_argv.extend(["--strategy", args.strategy])
-    if args.skip_nets:
-        sub_argv.extend(["--skip-nets", args.skip_nets])
-    if args.grid != 0.25:
-        sub_argv.extend(["--grid", str(args.grid)])
-    if args.trace_width != 0.2:
-        sub_argv.extend(["--trace-width", str(args.trace_width)])
-    if args.clearance != 0.15:
-        sub_argv.extend(["--clearance", str(args.clearance)])
-    if args.via_drill != 0.3:
-        sub_argv.extend(["--via-drill", str(args.via_drill)])
-    if args.via_diameter != 0.6:
-        sub_argv.extend(["--via-diameter", str(args.via_diameter)])
-    if args.mc_trials != 10:
-        sub_argv.extend(["--mc-trials", str(args.mc_trials)])
-    if args.iterations != 15:
-        sub_argv.extend(["--iterations", str(args.iterations)])
-    if args.verbose:
-        sub_argv.append("--verbose")
-    if args.dry_run:
-        sub_argv.append("--dry-run")
-    # Use command-level quiet or global quiet
-    if getattr(args, "quiet", False) or getattr(args, "global_quiet", False):
-        sub_argv.append("--quiet")
-    if getattr(args, "power_nets", None):
-        sub_argv.extend(["--power-nets", args.power_nets])
-    return route_main(sub_argv)
-
-
-def _run_reason_command(args) -> int:
-    """Handle reason command."""
-    from .reason_cmd import main as reason_main
-
-    sub_argv = [args.pcb]
-    if args.output:
-        sub_argv.extend(["-o", args.output])
-    if args.export_state:
-        sub_argv.append("--export-state")
-    if args.state_output:
-        sub_argv.extend(["--state-output", args.state_output])
-    if args.interactive:
-        sub_argv.append("--interactive")
-    if args.analyze:
-        sub_argv.append("--analyze")
-    if args.auto_route:
-        sub_argv.append("--auto-route")
-    if args.max_nets != 10:
-        sub_argv.extend(["--max-nets", str(args.max_nets)])
-    if args.drc:
-        sub_argv.extend(["--drc", args.drc])
-    if args.verbose:
-        sub_argv.append("--verbose")
-    if args.dry_run:
-        sub_argv.append("--dry-run")
-    return reason_main(sub_argv)
-
-
-def _run_placement_command(args) -> int:
-    """Handle placement command."""
-    if not args.placement_command:
-        print("Usage: kicad-tools placement <command> [options] <file>")
-        print("Commands: check, fix, optimize")
-        return 1
-
-    from .placement_cmd import main as placement_main
-
-    if args.placement_command == "check":
-        sub_argv = ["check", args.pcb]
-        if args.format != "table":
-            sub_argv.extend(["--format", args.format])
-        if args.pad_clearance != 0.1:
-            sub_argv.extend(["--pad-clearance", str(args.pad_clearance)])
-        if args.hole_clearance != 0.5:
-            sub_argv.extend(["--hole-clearance", str(args.hole_clearance)])
-        if args.edge_clearance != 0.3:
-            sub_argv.extend(["--edge-clearance", str(args.edge_clearance)])
-        if args.verbose:
-            sub_argv.append("--verbose")
-        # Use command-level quiet or global quiet
-        if getattr(args, "quiet", False) or getattr(args, "global_quiet", False):
-            sub_argv.append("--quiet")
-        return placement_main(sub_argv) or 0
-
-    elif args.placement_command == "fix":
-        sub_argv = ["fix", args.pcb]
-        if args.output:
-            sub_argv.extend(["-o", args.output])
-        if args.strategy != "spread":
-            sub_argv.extend(["--strategy", args.strategy])
-        if args.anchor:
-            sub_argv.extend(["--anchor", args.anchor])
-        if args.dry_run:
-            sub_argv.append("--dry-run")
-        if args.verbose:
-            sub_argv.append("--verbose")
-        # Use command-level quiet or global quiet
-        if getattr(args, "quiet", False) or getattr(args, "global_quiet", False):
-            sub_argv.append("--quiet")
-        return placement_main(sub_argv) or 0
-
-    elif args.placement_command == "optimize":
-        sub_argv = ["optimize", args.pcb]
-        if args.output:
-            sub_argv.extend(["-o", args.output])
-        if args.strategy != "force-directed":
-            sub_argv.extend(["--strategy", args.strategy])
-        if args.iterations != 1000:
-            sub_argv.extend(["--iterations", str(args.iterations)])
-        if args.generations != 100:
-            sub_argv.extend(["--generations", str(args.generations)])
-        if args.population != 50:
-            sub_argv.extend(["--population", str(args.population)])
-        if args.grid != 0.0:
-            sub_argv.extend(["--grid", str(args.grid)])
-        if args.fixed:
-            sub_argv.extend(["--fixed", args.fixed])
-        if args.dry_run:
-            sub_argv.append("--dry-run")
-        if args.verbose:
-            sub_argv.append("--verbose")
-        # Use command-level quiet or global quiet
-        if getattr(args, "quiet", False) or getattr(args, "global_quiet", False):
-            sub_argv.append("--quiet")
-        return placement_main(sub_argv) or 0
-
-    return 1
-
-
-def _run_optimize_command(args) -> int:
-    """Handle optimize-traces command."""
-    from .optimize_cmd import main as optimize_main
-
-    sub_argv = [args.pcb]
-    if args.output:
-        sub_argv.extend(["-o", args.output])
-    if args.net:
-        sub_argv.extend(["--net", args.net])
-    if args.no_merge:
-        sub_argv.append("--no-merge")
-    if args.no_zigzag:
-        sub_argv.append("--no-zigzag")
-    if args.no_45:
-        sub_argv.append("--no-45")
-    if args.chamfer_size != 0.5:
-        sub_argv.extend(["--chamfer-size", str(args.chamfer_size)])
-    if args.verbose:
-        sub_argv.append("--verbose")
-    if args.dry_run:
-        sub_argv.append("--dry-run")
-    # Use command-level quiet or global quiet
-    if getattr(args, "quiet", False) or getattr(args, "global_quiet", False):
-        sub_argv.append("--quiet")
-    return optimize_main(sub_argv)
-
-
-def _run_config_command(args) -> int:
-    """Handle config command."""
-    from .config_cmd import main as config_main
-
-    sub_argv = []
-    if args.show:
-        sub_argv.append("--show")
-    if args.init:
-        sub_argv.append("--init")
-    if args.paths:
-        sub_argv.append("--paths")
-    if args.user:
-        sub_argv.append("--user")
-    if args.config_action:
-        sub_argv.append(args.config_action)
-    if args.config_key:
-        sub_argv.append(args.config_key)
-    if args.config_value:
-        sub_argv.append(args.config_value)
-    return config_main(sub_argv) or 0
-
-
-def _run_interactive_command(args) -> int:
-    """Handle interactive command."""
-    from .interactive import main as interactive_main
-
-    sub_argv = []
-    if args.project:
-        sub_argv.extend(["--project", args.project])
-    return interactive_main(sub_argv)
-
-
-def _run_validate_command(args) -> int:
-    """Handle validate command."""
-    if not args.sync:
-        print("Usage: kicad-tools validate --sync [options] <project>")
-        print("Currently only --sync is supported.")
-        return 1
-
-    from .validate_sync_cmd import main as validate_sync_main
-
-    sub_argv = []
-    if args.validate_project:
-        sub_argv.append(args.validate_project)
-    if args.validate_schematic:
-        sub_argv.extend(["--schematic", args.validate_schematic])
-    if args.validate_pcb:
-        sub_argv.extend(["--pcb", args.validate_pcb])
-    if args.validate_format != "table":
-        sub_argv.extend(["--format", args.validate_format])
-    if args.validate_errors_only:
-        sub_argv.append("--errors-only")
-    if args.validate_strict:
-        sub_argv.append("--strict")
-    if args.validate_verbose:
-        sub_argv.append("--verbose")
-    return validate_sync_main(sub_argv)
+# Legacy standalone entry points
 
 
 def symbols_main() -> int:
